pyproject.toml
[project]
name = "job-cheat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.3",
    "firebase-admin>=6.5",
    "python-dotenv>=1.0",
    "selectolax>=0.3.34,<0.4",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.24.0", # ASGI 서버
    "gunicorn>=21.2.0", # WSGI/ASGI 서버
    "google-generativeai>=0.8.5",
    "numpy>=2.3.3",
    "scikit-learn>=1.7.2",
    "tiktoken>=0.11.0",
    "pinecone>=7.3.0",
]
